chore: Move contract code to own directory
