+++ databuilder/docs.py
@@ -2,7 +2,8 @@ import json
 import operator
 
 from .backends.base import BaseBackend
-from .concepts.tables import ClinicalEvents, PatientDemographics, PracticeRegistrations
+from .concepts.tables import ClinicalEvents, PracticeRegistrations
+from .contracts.tables import PatientDemographics
 
 
 def _build_backends():
