+++ tests/backends/test_base.py
@@ -1,10 +1,10 @@
 import pytest
 
 from databuilder.backends.base import BaseBackend, Column, MappedTable
-from databuilder.concepts import types
-from databuilder.concepts.table_contract import BackendContractError
-from databuilder.concepts.table_contract import Column as ColumnContract
-from databuilder.concepts.table_contract import TableContract
+from databuilder.contracts import types
+from databuilder.contracts.base import BackendContractError
+from databuilder.contracts.base import Column as ColumnContract
+from databuilder.contracts.base import TableContract
 from databuilder.query_engines.base_sql import BaseSQLQueryEngine
